# Api.Functions/openapi/v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rhema Bible Serverless API",
    "version": "1.0.0",
    "description": "Azure Functions HTTP API (`/api` prefix). Authenticated routes validate JWT and load the user from the database (soft-deleted users are rejected). JSON bodies use **camelCase** property names. Enums serialize as **PascalCase** strings (`JsonStringEnumConverter` without a naming policy), matching `Api.Functions/Program.cs` worker JSON options."
  },
  "servers": [{ "url": "/api", "description": "Functions HTTP route prefix" }],
...
  "paths": {
...
    "/v1/admin/profile": {
      "get": {
        "tags": ["Admin"],
        "summary": "Admin profile (app user if registered, else Clerk claims only)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" }, "401": { "description": "Unauthorized" }, "403": { "description": "Not admin" } }
      }
    },
    "/v1/admin/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "Paged users",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Paged users" } }
      }
    },
    "/v1/admin/users/{userId}": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get user",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not found" } }
      }
    },
    "/v1/admin/users/{userId}/activate": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Activate user",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/v1/admin/users/{userId}/deactivate": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Deactivate user",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/v1/admin/users/{userId}/subscription": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Update subscription / plan",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateSubscriptionDto" } } }
        },
        "responses": { "200": { "description": "OK" }, "400": { "description": "Invalid body" } }
      }
    },
    "/v1/admin/dashboard": {
      "get": {
        "tags": ["Admin"],
        "summary": "Dashboard analytics",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/v1/admin/dashboard/statistics": {
      "get": {
        "tags": ["Admin"],
        "summary": "Dashboard statistics (JSON or CSV via `format=csv`)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "JSON export or CSV attachment",
            "content": {
              "application/json": { "schema": { "type": "object" } },
              "text/csv": { "schema": { "type": "string", "format": "binary" } }
            }
          }
        }
      }
    },
    "/v1/admin/users/{userId}/ai-quota": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get user AI quota",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/v1/admin/users/{userId}/ai-quota/reset": {
      "post": {
        "tags": ["Admin"],
        "summary": "Reset user AI quota",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/v1/admin/users/{userId}/ai-quota/set-remaining": {
      "post": {
        "tags": ["Admin"],
        "summary": "Set remaining AI quota (query: `remainingThisMonth`)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } },
          {
            "name": "remainingThisMonth",
            "in": "query",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Remaining AI units for the month"
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/v1/admin/account-deletion/purge": {
      "post": {
        "tags": ["Admin"],
        "summary": "Purge soft-deleted users past grace period",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Purge summary" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from login/register/refresh."
      },
      "revenueCatWebhookBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Raw RevenueCat webhook secret as Bearer token."
      }
    },
    "schemas": {
...
      "SubscriptionType": {
        "type": "string",
        "enum": ["Free", "Premium", "PremiumMonthly", "PremiumYearly"],
        "description": "Domain enum `SubscriptionType` serialized as string."
      },
      "UpdateSubscriptionDto": {
        "type": "object",
        "required": ["subscriptionType"],
        "properties": {
          "subscriptionType": { "$ref": "#/components/schemas/SubscriptionType" },
          "subscriptionExpiresAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
    }
  }
}